+++ projects/packages/protect-status/src/class-status.php
@@ -14,7 +14,7 @@ use Automattic\Jetpack\Protect_Models\Status_Model;
  */
 class Status {
 
-	const PACKAGE_VERSION = '0.5.1';
+	const PACKAGE_VERSION = '0.5.2';
 	/**
 	 * Name of the option where status is stored
 	 *