+++ projects/plugins/jetpack/tests/php/lib/class-wp-test-jetpack-rest-testcase.php
@@ -10,7 +10,7 @@ if ( defined( 'IS_WPCOM' ) && IS_WPCOM ) {
 	require_once dirname( __DIR__, 3 ) . '/core/includes/testcase-rest-api.php';
 }
 
-abstract class WP_Test_Jetpack_REST_Testcase extends WP_Test_REST_TestCase {
+abstract class WP_Test_Jetpack_REST_TestCase extends WP_Test_REST_TestCase {
 	protected $server;
 
 	/**