+++ projects/packages/assets/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.8] - 2025-02-24
+### Added
+- User data: Add permissions to the current user object [#41859]
+
 ## [4.0.7] - 2025-02-17
 ### Changed
 - Update dependencies.
@@ -564,6 +568,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Statically access asset tools
 
+[4.0.8]: https://github.com/Automattic/jetpack-assets/compare/v4.0.7...v4.0.8
 [4.0.7]: https://github.com/Automattic/jetpack-assets/compare/v4.0.6...v4.0.7
 [4.0.6]: https://github.com/Automattic/jetpack-assets/compare/v4.0.5...v4.0.6
 [4.0.5]: https://github.com/Automattic/jetpack-assets/compare/v4.0.4...v4.0.5