+++ projects/packages/protect-status/CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2025-02-24
+### Changed
+- Update dependencies.
+
 ## [0.5.1] - 2025-02-11
 ### Fixed
-- Protect Status: ensure vulnerabilities property is always an array. [#41694]
+- Protect Status: Ensure vulnerabilities property is always an array. [#41694]
 
 ## [0.5.0] - 2025-02-10
 ### Changed
@@ -80,6 +84,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#37894]
 
+[0.5.2]: https://github.com/Automattic/jetpack-protect-status/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/Automattic/jetpack-protect-status/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/Automattic/jetpack-protect-status/compare/v0.4.3...v0.5.0
 [0.4.3]: https://github.com/Automattic/jetpack-protect-status/compare/v0.4.2...v0.4.3