+++ projects/packages/redirect/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.2] - 2025-02-24
+### Changed
+- Update dependencies.
+
 ## [3.0.1] - 2024-11-25
 ### Changed
 - Updated package dependencies. [#40258]
@@ -221,6 +225,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Create Jetpack Redirect package
 
+[3.0.2]: https://github.com/Automattic/jetpack-redirect/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/Automattic/jetpack-redirect/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/Automattic/jetpack-redirect/compare/v2.0.5...v3.0.0
 [2.0.5]: https://github.com/Automattic/jetpack-redirect/compare/v2.0.4...v2.0.5